# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sierra Sync API
  description: |
    Real-time lead and sales data synchronization platform between Sierra Interactive and popular CRMs.
    
    ## Authentication
    The API uses JWT Bearer tokens for authentication. Obtain a token via the `/auth/login` endpoint.
    
    ## Rate Limiting
    - Free tier: 100 requests/minute
    - Starter: 500 requests/minute
    - Professional: 2000 requests/minute
    - Enterprise: 10000 requests/minute
    
    ## Webhooks
    Register webhooks to receive real-time updates on lead and sync events.
  version: 1.0.0
  contact:
    name: Sierra Sync Support
    email: support@sierrasync.com
    url: https://sierrasync.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sierrasync.com/v1
    description: Production server
  - url: https://staging-api.sierrasync.com/v1
    description: Staging server
  - url: http://localhost:3001/v1
    description: Local development
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login to account
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        organizationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'